Unify types: bint block types, double transport, ACC internal, BINT typedefs
- DUT types renamed: int8/int16/int32 → bint8/bint16/bint32 (128-bit block)
- External transport unified to double (no more float32)
- ACC format fully internal to dsp_*.h (Python never sees q12_22_t)
- Golden C params use BINT16/BINT32 typedefs
- dsp_*.h wrappers use const_cast CC() macro for golden C compat
- Restructured: golden_c/current/include/ (hardware), golden_c/dsp/ (adapter)
- auto_register handles bint/int/double token mapping
- dut_to_double binding for comparison path
- LayerNorm framework with ComputeType parameter

+++ projects/dsp-core/src/dsp/golden/call.py
@@ -50,7 +50,7 @@ def convert(data: np.ndarray, src_type: str, dst_type: str) -> np.ndarray:
             f"C 函数 '{func_name}' 在 .so 中不存在。"
         )
     dst = np.empty_like(data)
-    func(data.astype(np.float32), dst, data.size)
+    func(data.astype(np.float64), dst, data.size)
     return dst
 
 
@@ -76,7 +76,7 @@ def compute(*inputs: np.ndarray, query: ComputeKey) -> dict:
     from .op_convention import require_convention
     conv = require_convention(op)
 
-    inputs_f32 = [inp.astype(np.float32) for inp in inputs]
+    inputs_f32 = [inp.astype(np.float64) for inp in inputs]
     key = info["key"]
     out_np = conv.call_c_func(func, *inputs_f32, compute_key=key)
     return {
